added main database
the name leaderboard is a misnomer. This database contains leaderboards for all three levels (leaderboard_lvl1, leaderboard_lvl2, leaderboard_lvl3) as well as a table to track level progress (level_progress). For each leaderboard, there are columns for Name, Score, and Time. For the level progress table, theres columns for Name and Current Level. The default value for current level is set to 0 which corresponds with starting the game from scratch.