Able to run "/" test! Todo:fix error, implement+test rest endpoints
rest as in the remaining ones. i.e. create_user, load_users.

The current error just is a NameError for a constant I created with the number of levels a while ago but deleted. It's currently in use in the `SQL_FOR_CREATE_USERS` which may not be necessary at all...
**Make sure to verify if schema.sql is executed**
- i.e. try to verify the tables were created after the home endpoint is loaded.