Add app calls to init db! Todo: determine why missing print msg.
Learned from the flask documentation that you usually import a function called `init_app` from db.py which basically just calls `init_db`. Currently when I run the test for the "/" endpoint, I don't see a print message for "Initialized database!".

However, it may just be a `flask run` thing. None of the print statements are loading into there, which is why I've been using `... pytest -s` to show the stdout.

I'm guessing since this code to initialize the database isn't directly called in an endpoint, its not redirected to the pytest.

Todo: try to create a user!