Wrote add_user(username) + `flask drop-db` (todo: endpoint for add).
**drop-db**
Wrote a click command for drop-db so that as we change the schema we can re-initialize the table. i.e. `flask drop-db && flask init-db`. Currently I only modified the schema to set a default user_id so I don't have to specify it and it hopefully auto-increments from there since its a primary key.

**add_user**
I tested the sql command with a sample username and it works in the sqlite3 command line. I have not tested it through flask i.e. an HTTP request as I have not implemented the endpoint for it.

Considering if I want to validate the username before we enter it or we just let the SQL integrity constraint do the work.