db.incr_score() works! load_leaderboard fails.
Tested `db.increment_score()` and `db.initialize_score()` in flask shell, and they work!

Updated add_user() to initialize the row in the leaderboard so that I wouldn't have to look the user up in that table to give them a leaderboard row with the same id. Todo: verify (i.e. after fixing load_leaderboard()

 I need to change how I loop through that table to get the attributes I want (debug and see what is printed out). Lots of cleaning up to do.