[POST /increment_score] PASS tests :) Note set usr
See top of `conftest.py` for an important message about running this test!
Basically, I have the test create a fresh user every time, so you gotta set the username at the top of this file to make it not complain about that username already existing or it not starting from a fresh state (i.e. level 1).

The test checks if we can go from level 1 to 2. I didn't explicitly test 2 -> 3 and 3 staying at three.